requirements.txt
selenium==4.35.0
Pillow==9.5.0
fpdf==1.7.2
webdriver-manager==4.0.2
